implemented Kafka and Unit tests, code refactoring

+++ CarService/Controllers/TyreController.cs
@@ -1,4 +1,5 @@
-﻿using CarService.BL.Interfaces;
+﻿using System.Threading;
+using CarService.BL.Interfaces;
 using CarService.BL.Services;
 using CarService.Models.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -20,11 +21,19 @@ namespace CarService.Controllers
             _logger = logger;
         }
 
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPost(nameof(AddTyre))]
         public async Task<IActionResult> AddTyre([FromBody] Tyre tyre)
         {
+            var result = await _tyreService.GetTyreByName(tyre.TyreName);
+
+            if (result != null)
+            {
+                return NotFound("Tyre already exist");
+            }
+
             return Ok(await _tyreService.AddTyre(tyre));
         }
 
@@ -33,6 +42,11 @@ namespace CarService.Controllers
         public async Task<IActionResult> GetAllTyres()
         {
             var result = await _tyreService.GetAllTyres();
+
+            if (result.Count() <= 0)
+            {
+                return BadRequest("There aren't any tyres in the collection");
+            }
 
             return Ok(result);
         }
@@ -47,7 +61,7 @@ namespace CarService.Controllers
 
             var result = await _tyreService.GetTyreById(tyreId);
 
-            if (result == null) return NotFound($"Tyre not exist");
+            if (result == null) return NotFound($"Tyre does not exist");
 
             return Ok(result);
         }
@@ -56,22 +70,47 @@ namespace CarService.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPut(nameof(UpdateTyre))]
-        public async Task<IActionResult> UpdateTyre(Tyre tyre)
+        public async Task<IActionResult> UpdateTyre(Tyre tyre, int id)
         {
-            return Ok(await _tyreService.UpdateTyre(tyre));
+            if (id <= 0) return BadRequest($"Parameter id: {id} must be greater than zero");
+
+            var tyreToUpdate = await _tyreService.GetTyreById(id);
+
+            if (tyreToUpdate == null) return NotFound($"Tyre with id:{id} does not exist");
+
+            return Ok(await _tyreService.UpdateTyre(tyre, id));
         }
 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [HttpDelete(nameof(DeleteTyre))]
         public async Task<IActionResult> DeleteTyre(int tyreId)
         {
+            if (tyreId <= 0) return BadRequest($"Parameter id: {tyreId} must be greater than zero");
+
             if (await _tyreService.GetTyreById(tyreId) == null)
             {
-                return BadRequest("Client not exist");
+                return NotFound($"Tyre does not exist");
             }
 
             var result = await _tyreService.DeleteTyre(tyreId);
+
+            return Ok(result);
+        }
+
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [HttpGet(nameof(GetTyreByName))]
+        public async Task<IActionResult> GetTyreByName(string tyreName)
+        {
+            var result = await _tyreService.GetTyreByName(tyreName);
+
+            if (result == null)
+            {
+                return NotFound($"Tyre does not exist");
+            }
 
             return Ok(result);
         }
